+++ pyobs/modules/flatfield/flatfield.py
@@ -84,7 +84,7 @@ class FlatField(Module, IFlatField):
                             filter=filter, binning=binning)
 
     @timeout(3600)
-    def flat_field(self, filter_name: str, count: int = 20, binning: int = 1, *args, **kwargs) -> Tuple[int, int]:
+    def flat_field(self, filter_name: str, count: int = 20, binning: int = 1, *args, **kwargs) -> Tuple[int, float]:
         """Do a series of flat fields in the given filter.
 
         Args:
